% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mean.R
\name{calculate_mean}
\alias{calculate_mean}
\title{Calculate the Mean of a Numeric Vector}
\usage{
calculate_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the vector \code{x}.
}
\description{
This function calculates the mean of a numeric vector. The mean is defined as:
\deqn{\mu = \frac{1}{n} \sum_{i=1}^{n} x_i}
}
\details{
The mean \eqn{\mu} is calculated using the formula:
\deqn{\mu = \frac{1}{n} \sum_{i=1}^{n} x_i}
where \eqn{x_i} represents each element of the vector and \eqn{n} is the length of the vector.
}
\examples{
x <- c(1, 2, 3, 4, 5)
calculate_mean(x)

}
