% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IApimat.R
\docType{data}
\name{IApimat}
\alias{IApimat}
\title{A matrix used for variance estimation in IAdata}
\description{
The matrix that is used for variance estimation in IAdata. The sample is collected from a straitified random sampling.
}
\examples{
data(IApimat) 
}
\keyword{matrix}
