% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEcalib.R
\name{GEcalib}
\alias{GEcalib}
\title{Generalized Entropy Calibration}
\usage{
GEcalib(
  Xs,
  total,
  d = NULL,
  entropy = c("SL", "EL", "ET", "CE", "HD", "PH"),
  method = "Newton",
  control = list(maxit = 1e+05, allowSingular = T)
)
}
\arguments{
\item{Xs}{A matrix of auxiliary variables.}

\item{total}{A vector of population totals.}

\item{d}{An optional vector of initial weights.}

\item{entropy}{An entropy function(Squared-loss:"SL", Empirical Likelihood: "EL", Exponential Tilting: "ET", Cross-Entropy: "CE", Helinger Distance: "HD", "PH")}

\item{method}{an optional vector that can be used in \code{nleqslv}.}

\item{control}{an optional list that can be used in \code{nleqslv}.}
}
\value{
A vector of calibration weights.
}
\description{
\code{GEcalib} computes the generalized entropy calibration weight proposed by Kwon et.al. (2024).
The \code{GEcalib} weight minimizes the negative generalized entropy:
\deqn{\sum_{i \in A} G(\omega_i)}
subject to the calibration constraints \eqn{\sum_{i \in A} \omega_i \bm z_i = \sum_{i \in U} \bm z_i},
where \eqn{A} is the index of sample, \eqn{U} is the index of population, and
\eqn{\bm z_i} are the auxiliary variables whose population totals are known.
}
\examples{
Xs=cbind(
  c(1,1,1,1,1,1,1,1,1,1),
  c(1,1,1,1,1,0,0,0,0,0),
  c(1,2,3,4,5,6,7,8,9,10)
)
# inclusion probabilities
piks=rep(0.2,times=10); d=1/piks
# vector of population totals
total=c(50,24,290)

# Calibration weights
# calib(Xs, total, d = d, method="raking") * d
GEcalib(Xs, total, d, entropy = "ET")
GEcalib(Xs, total, entropy = "ET")

# calib(Xs, total, d = d, method="linear") * d
GEcalib(Xs, total, d, entropy = "SL")
GEcalib(Xs, total, entropy = "SL")
}
\references{
Kwon, Y., Kim, J., & Qiu, Y. (2024). Debiased calibration estimation using generalized entropy in survey sampling.
Arxiv preprint <https://arxiv.org/abs/2404.01076>

Deville, J. C., and Särndal, C. E. (1992). Calibration estimators in survey sampling.
Journal of the American statistical Association, 87(418), 376-382.
}
