% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEcalib.R
\name{GEcalib}
\alias{GEcalib}
\title{Generalized Entropy Calibration}
\usage{
GEcalib(
  formula,
  dweight,
  data = NULL,
  const,
  method = c("GEC", "GEC0", "DS"),
  entropy = c("SL", "EL", "ET", "CE", "HD", "PH"),
  weight.scale = 1,
  G.scale = 1,
  del = NULL,
  K_alpha = NULL,
  is.total = T
)
}
\arguments{
\item{method}{an optional vector that can be used in \code{nleqslv}.}

\item{entropy}{An entropy function(Squared-loss:"SL", Empirical Likelihood: "EL", Exponential Tilting: "ET", Cross-Entropy: "CE", Helinger Distance: "HD", "PH")}

\item{Xs}{A matrix of auxiliary variables.}

\item{total}{A vector of population totals.}

\item{d}{An optional vector of initial weights.}

\item{control}{an optional list that can be used in \code{nleqslv}.}
}
\value{
A vector of calibration weights.
}
\description{
\code{GEcalib} computes the generalized entropy calibration weight proposed by Kwon et.al.(2024).
The \code{GEcalib} weight minimizes the negative generalized entropy:
\deqn{\sum_{i \in A} G(\omega_i)}
subject to the calibration constraints \eqn{\sum_{i \in A} \omega_i \bm z_i = \sum_{i \in U} \bm z_i},
where \eqn{A} is the index of sample, \eqn{U} is the index of population, and
\eqn{\bm z_i^T = (\bm x_i^T, g(d_i))} are the auxiliary variables whose population totals are known.
}
\section{Summary}{


\tabular{cc}{
\strong{GEC} \tab \strong{DS} \cr
\eqn{\min_{\bm \omega} \left(-H(\bm \omega)\right) = \sum_{i \in A}G(\omega_i) \quad} \tab 
\eqn{\quad \min_{\bm \omega} D(\bm \omega, \bm d) = \sum_{i \in A}d_iG(\omega_i / d_i)} \cr
\deqn{G(\omega) = \begin{cases} \frac{1}{r(r+1)} \omega^{r+1} & r \neq 0, -1\\ 
\omega \log \omega - \omega & r = 0\text{(ET)} \\ 
-\log \omega & r = -1\text{(EL)} \end{cases}} 
\tab \deqn{G(\omega) = \begin{cases} \frac{1}{r(r+1)} \left(\omega^{r+1} - (r+1)\omega + r\right) & r \neq 0, -1 \\
\omega \log \omega - \omega + 1 & r = 0\text{(ET)} \\
-\log \omega + \omega - 1 & r = -1\text{(EL)} \end{cases}} \cr
}
}

\examples{
Xs=cbind(
  c(1,1,1,1,1,1,1,1,1,1),
  c(1,1,1,1,1,0,0,0,0,0),
  c(1,2,3,4,5,6,7,8,9,10)
)
# inclusion probabilities
piks=rep(0.2,times=10); d=1/piks
# vector of population totals
total=c(50,24,290)

# Calibration weights
# calib(Xs, total, d = d, method="raking") * d
GEcalib(Xs, total, d, entropy = "ET")
GEcalib(Xs, total, entropy = "ET")

# calib(Xs, total, d = d, method="linear") * d
GEcalib(Xs, total, d, entropy = "SL")
GEcalib(Xs, total, entropy = "SL")
}
\references{
Kwon, Y., Kim, J., & Qiu, Y. (2024). Debiased calibration estimation using generalized entropy in survey sampling.
Arxiv preprint <https://arxiv.org/abs/2404.01076>

Deville, J. C., and Särndal, C. E. (1992). Calibration estimators in survey sampling.
Journal of the American statistical Association, 87(418), 376-382.
}
